improve user facing exception messages
